Heroku FastAPI/requirements.txt
fastapi
gunicorn
jsonpickle
numpy
opencv-python
Pillow
tensorflow
uvicorn